+++ ohsome_quality_api/indicators/roads_thematic_accuracy/queries/oneway.sql
@@ -5,28 +5,28 @@ WITH bpoly AS (
 )
 
 select
-    SUM(dlm_length) as total_dlm,
+    SUM(matched_length) + SUM(not_matched_length) as total_dlm,
         SUM(
             CASE
-                WHEN oneway is not NULL AND far is not NULL THEN dlm_length
+                WHEN oneway is not NULL AND far is not NULL THEN matched_length
                 ELSE 0
             END
         ) AS present_in_both,
         SUM(
         CASE
-            WHEN oneway is not NULL AND far is NULL THEN dlm_length
+            WHEN oneway is not NULL AND far is NULL THEN matched_length
             ELSE 0
         END
     ) AS only_osm,
 		SUM(
 			CASE
-				WHEN oneway IS NULL AND far IS NOT NULL THEN dlm_length
+				WHEN oneway IS NULL AND far IS NOT NULL THEN matched_length
 				ELSE 0
 			end
 		) AS only_dlm,
         SUM(
         CASE
-            WHEN oneway is NULL AND far is NULL THEN dlm_length
+            WHEN oneway is NULL AND far is NULL THEN matched_length
             ELSE 0
         END
         ) AS missing_both,
@@ -35,7 +35,7 @@ select
                 WHEN oneway is not NULL
                      AND far is not NULL
                      AND oneway = far
-                     AND ((angle_osm > 0 AND angle_dlm > 0) OR (angle_osm < 0 AND angle_dlm < 0)) THEN dlm_length
+                     AND ((angle_osm > 0 AND angle_dlm > 0) OR (angle_osm < 0 AND angle_dlm < 0)) THEN matched_length
                 ELSE 0
             END
         ) AS present_in_both_agree,
@@ -45,16 +45,11 @@ select
                      AND far is not NULL
                      AND (oneway != far
                      OR ((angle_osm < 0 AND angle_dlm > 0) OR (angle_osm < 0 AND angle_dlm > 0))
-                     ) THEN dlm_length
+                     ) THEN matched_length
                 ELSE 0
             END
         ) AS present_in_both_not_agree,
-        SUM(
-            CASE
-                WHEN osm_id IS NULL THEN dlm_length
-                ELSE 0
-            END
-        ) AS not_matched
+        SUM(not_matched_length) AS not_matched
 FROM road_thematic_accuracy rta, bpoly b
 WHERE
     ST_Intersects(rta.geom, b.geometry);